Implemented the Events DAO. A large amount of the code was spent creating the actual DAO object since many parameters needed to be parsed to the appropriate type and a specific event factory's create method needed to be called by identifying the type of event in the CSV.
In the CSV file, any unnecessary parameters for an event are an empty string for that parameters row except for peopleInvited. This is because the peopleInvited parameter is the only parameter that can identify an invite only event and invite only events can have no one invited. Thus, having it as an empty string in the CSV file would not be appropriate, so I set it to be -1 for non invite only events.

Also, I changed the event owner parameter to being the owner ID rather than an actual owner object. This would make it easier to read from the CSV file when creating an event as we can just read an owner ID rather than several parameter, which wouldn't be needed since we aren't creating another user with those.