Unrelated to the use case - Updated the types for various instances in the entities, such as event time which was switched from a string to LocalDateTime and switching the identifier for user from the ID to username (int to string).
For the create event use case, all the code for the application business rules was implemented including various test cases for create event (which all pass), such as detecting valid/invalid inputs, adding events to the DAO, and editing the organizer's created events instance.

Depending on how the front end is implemented, certain aspects of the code might change in terms of parsing the inputs as I don't know how we are going to take in each input yet.